+++ CHANGELOG.md
@@ -3,7 +3,8 @@ All notable changes to this project will be documented in this file.
 
 ## v0.2.2 - ??
 ### Added
- - Resume edit, create
+ - Resumes: edit, create
+ - Vacancies: favorited,
 
 ## v0.2.1 - 2017-04-24
 ### Changed: